# server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Arth API",
    "description": "Realtime paper-trading broker API (SSE market stream, session cookie auth).",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "http://127.0.0.1:4000", "description": "Local" }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "arth_session"
      }
    },
...
  },
  "paths": {
...
    "/api/orders": {
      "get": {
        "summary": "List orders",
        "tags": ["Orders"],
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "200": { "description": "Orders for current user" },
          "401": { "description": "Not authenticated" },
          "429": { "description": "Rate limited" }
        }
      },
      "post": {
        "summary": "Place order",
        "tags": ["Orders"],
        "security": [{ "cookieAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "symbol": { "type": "string" },
                  "side": { "type": "string", "enum": ["BUY", "SELL"] },
                  "qty": { "type": "number" },
                  "type": { "type": "string" },
                  "price": { "type": "number" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Order accepted" },
          "201": { "description": "Order created" },
          "400": { "description": "Validation error" },
          "401": { "description": "Not authenticated" },
          "429": { "description": "Rate limited" }
        }
      }
    },
    "/api/orders/preview": {
      "post": {
        "summary": "Preview order charges",
        "tags": ["Orders"],
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "200": { "description": "Preview" }
        }
      }
    },
    "/api/portfolio/summary": {
      "get": {
        "summary": "Portfolio summary",
        "tags": ["Portfolio"],
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "200": { "description": "Cash + equity snapshot" }
        }
      }
    },
    "/api/portfolio/holdings": {
      "get": {
        "summary": "Holdings",
        "tags": ["Portfolio"],
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "200": { "description": "Position list" }
        }
      }
    },
    "/api/portfolio/funds": {
      "get": {
        "summary": "Funds / cash",
        "tags": ["Portfolio"],
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "200": { "description": "Available funds" }
        }
      }
    },
...
  }
}